Fix Dockerfile so beta realease will be from beta branch

+++ threadfin.go
@@ -6,7 +6,7 @@
 package main
 
 import (
-	"bytes"
+	"bufio"
 	"flag"
 	"fmt"
 	"os"
@@ -42,13 +42,13 @@ var GitHub = GitHubStruct{Branch: "Main", User: "marcelGoerentz", Repo: "Threadf
 const Name = "Threadfin"
 
 // Version : Version, die Build Nummer wird in der main func geparst.
-const Version = "1.2.4-beta"
+const Version = "1.2.5-beta"
 
 // DBVersion : Datanbank Version
 const DBVersion = "0.5.0"
 
 // APIVersion : API Version
-const APIVersion = "1.2.4-beta"
+const APIVersion = "1.2.5-beta"
 
 var homeDirectory = fmt.Sprintf("%s%s.%s%s", src.GetUserHomeDirectory(), string(os.PathSeparator), strings.ToLower(Name), string(os.PathSeparator))
 var samplePath = fmt.Sprintf("%spath%sto%sthreadfin%s", string(os.PathSeparator), string(os.PathSeparator), string(os.PathSeparator), string(os.PathSeparator))
@@ -151,25 +151,22 @@ func main() {
 	// Https Webserver
 	system.Flag.UseHttps = *useHttps
 
-	// kill all ffmpeg and VLC processess
-	cmdFindFFmpeg := "pgrep -f 'ffmpeg.*Title=Threadfin'"
-	cmdFindVLC := "pgrep -f 'cvlc.*meta-title=Threadfin'"
 
-	// Execute the commands
-	ffmpegPIDs, _ := getPIDs(cmdFindFFmpeg)
-
-	vlcPIDs, _ := getPIDs(cmdFindVLC)
-
-	// Combine PIDs into one slice
-	allPIDs := append(ffmpegPIDs, vlcPIDs...)
-
-	// Kill all the processes by PID
-	for _, pid := range allPIDs {
-		err := killProcess(pid)
-		if err != nil {
-			fmt.Printf("Error killing process %s: %v", pid, err)
-		} else {
-			fmt.Printf("Successfully killed process %s", pid)
+	// Kill all remaining processes and remove PIDs file
+	pids, err := getPIDsFromFile(*system)
+	if err != nil {
+		fmt.Printf("Error scanning file PIDs: %v", err)
+	} else {
+		if len(pids) > 0 {
+			for _, pid := range pids {
+				err := killProcess(pid)
+				if err != nil {
+					fmt.Printf("Error killing process %s: %v", pid, err)
+				} else {
+					fmt.Printf("Successfully killed process %s", pid)
+				}
+			}
+			os.Remove(system.Folder.Temp + "PIDs")
 		}
 	}
 
@@ -210,7 +207,7 @@ func main() {
 		os.Exit(0)
 	}
 
-	err := src.Init()
+	err = src.Init()
 	if err != nil {
 		src.ShowError(err, 0)
 		os.Exit(0)
@@ -241,16 +238,28 @@ func main() {
 
 }
 
-func getPIDs(command string) ([]string, error) {
-	var out bytes.Buffer
-	cmd := exec.Command("bash", "-c", command)
-	cmd.Stdout = &out
-	err := cmd.Run()
-	if err != nil {
-		return nil, err
+func getPIDsFromFile(system src.SystemStruct) ([]string, error){
+	var err error
+	pids := []string{}
+	// Open the file
+	pidsFile := system.Folder.Temp + "PIDs"
+	_, err_stat := os.Stat(pidsFile)
+	if os.IsExist(err_stat) {
+		var file *os.File
+		file, err = os.Open(pidsFile)
+		defer file.Close() // Close the file when done
+
+		// Create a scanner
+		scanner := bufio.NewScanner(file)
+
+		// Read line by line
+		for scanner.Scan() {
+			line := scanner.Text()
+			pids = append(pids, line)
+		}
 	}
-	pids := strings.Fields(out.String())
-	return pids, nil
+	return pids, err
+
 }
 
 // killProcess kills a process by its PID
